+++ src/apis/profile/uploadImage.ts
@@ -1,5 +1,5 @@
 import { UPDATE_PROFILE_IMAGE_PATH } from '@/utils/api_paths'
-import axiosInstance from './api'
+import axiosInstance from '../api'
 
 async function uploadImage(formData: FormData) {
   try {
