+++ src/pages/Profile/components/profile-edit-drawer/UploadImageButton.tsx
@@ -3,7 +3,7 @@ import { ChangeEvent, useCallback, useRef } from 'react'
 import tw, { styled } from 'twin.macro'
 
 import handleImageFormData from '@/utils/handleImageFormData'
-import uploadImage from '@/apis/uploadImage'
+import uploadImage from '@/apis/profile/uploadImage'
 import { useProfileStore } from '@/stores/userProfileStore'
 
 const FileInput = styled.input`
